# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Be
  description: |
    The API provides CSR backend API.
basePath: /api/
schemes:
  - https
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /equipment/status_names:
    post:
      summary: Creating a new equipment status.
      security:
        - Bearer: [ ]
      tags:
        - EquipmentStatusName
      operationId: postEquipmentStatusName
      parameters:
        - name: name
          in: body
          description: Status Name for equipment
          required: true
          schema:
            $ref: "#/definitions/NewEquipmentStatusName"
      responses:
        201:
          description: Status name for equipment has been created
          schema:
            $ref: "#/definitions/SuccessEquipmentStatusNameOperationResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
    get:
      summary: List all statuses.
      security:
        - Bearer: [ ]
      tags:
        - EquipmentStatusName
      operationId: ListEquipmentStatusNames
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/ListEquipmentStatusNames"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
...
  /pet_size/{petSizeId}:
    parameters:
      - name: petSizeId
        in: path
        required: true
        description: pet size id
        type: integer
    get:
      summary: Get pet size by id
      security:
        - Bearer: [ ]
      tags:
        - PetSize
      operationId: GetPetSize
      responses:
        200:
          description: Pet size has been found
          schema:
            $ref: "#/definitions/PetSizeResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
    delete:
      summary: Delete pet size by id
      security:
        - Bearer: [ ]
      tags:
        - PetSize
      operationId: DeletePetSize
      responses:
        200:
          description: Pet size has been deleted
          schema:
            type: string
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
    put:
      summary: Edit pet size by id
      security:
        - Bearer: [ ]
      tags:
        - PetSize
      operationId: EditPetSize
      parameters:
        - name: EditPetSize
          in: body
          description: Edit an pet size
          required: true
          schema:
            $ref: "#/definitions/PetSize"
      responses:
        200:
          description: Pet size edited
          schema:
            $ref: "#/definitions/PetSizeResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  Error:
    type: object
    required:
      - data
    properties:
      data:
        type: object
        properties:
          correlationId:
            type: string
          message:
            type: string
...
  EquipmentStatusName:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
      name:
        type: string
  NewEquipmentStatusName:
    type: object
    required:
      - name
    properties:
      name:
        type: string
  ListEquipmentStatusNames:
    type: array
    items:
      $ref: "#/definitions/EquipmentStatusName"
  SuccessEquipmentStatusNameOperationResponse:
    type: object
    required:
      - data
    properties:
      data:
        $ref: "#/definitions/EquipmentStatusName"
...
  PetSize:
    type: object
    required:
      - name
      - size
    properties:
      name:
        type: string
        example: "big"
      size:
        type: string
        example: "76x48x53"
      is_universal:
        type: boolean
        example: false
...
  PetSizeResponse:
    type: object
    required:
      - id
      - name
      - size
    properties:
      id:
        type: integer
      name:
        type: string
      size:
        type: string
      is_universal:
        type: boolean

  #Active Areas
...